+++ android/app/src/main/kotlin/com/example/obs_for_sama/MainActivity.kt
@@ -2,4 +2,4 @@ package com.example.obs_for_sama
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()