+++ lib/core/themes/app_bar_theme.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+import 'package:obs_for_sama/core/core_theme_index.dart';
+
+AppBarTheme kAppBarTheme = AppBarTheme(
+  backgroundColor: kPrimaryColor,
+  titleTextStyle: kheadlineLarge,
+);